+++ package.json
@@ -6,12 +6,16 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "faker": "^6.6.6",
+    "gridjs": "^5.1.0",
+    "gridjs-react": "^5.1.0",
     "howl": "^0.2.0",
     "howler": "^2.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "react-table": "^7.8.0",
     "styled-components": "^5.3.5",
     "tone": "^14.7.77"
   },
@@ -38,5 +42,19 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "compilerOptions": {
+    "baseUrl": "src",
+    "paths": {
+      "*": [
+        "*"
+      ]
+    }
+  },
+  "include": [
+    "src"
+  ],
+  "devDependencies": {
+    "@faker-js/faker": "^7.6.0"
   }
 }
