+++ src/components/ExercisesEditor/styles.js
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const StyledUl = styled.ul`
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  border-bottom: 2px solid black;
+  width: 100%;
+`;
+export const StyledLi = styled.li`
+  padding: 0;
+  margin: 0;
+  font-family: Arial, Helvetica, sans-serif;
+  display: flex;
+`;
+export const StyledDataContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
