+++ src/utils/constants/ACTIONS.js
@@ -1,0 +1,17 @@
+
+export const ACTIONS = {
+    DELETE_RHYTEM: "delete",
+    UPDATE_TEMPO: "update tempo",
+    UPDATE_CURR_EX_NAME: "update exercise name",
+    UPDATE_CURR_EX_TYPE: "update exercise type",
+    UPDATE_CURR_EX_TEMPO: "update exercise tempo",
+    GET_LOCAL_STORAGE: "get local storage",
+    SET_LOCAL_STORAGE: "set local storage",
+    ADD_EXERCISE: "add exercise",
+    SET_EDIT_ON: "set edit on",
+    SET_EDIT_OFF: "set edit off",
+    SET_CURR_LIST_ITEM_INDEX: "set current list item index",
+    FILL_INPUTS_WITH_CURR_EX: "fill inputs with current exercise",
+    EDIT_EXERCISE: "edit exercise",
+    DELETE_EXERCISE: "delete exercise",
+  };
