+++ src/utils/constants/initialState.js
@@ -1,0 +1,11 @@
+
+
+
+export const initialState = {
+    exList: [], // [{name: string, tempo:number, type:string}]
+    currExType: "",
+    currExTempo: 0,
+    currExName: "",
+    isEditOn: false,
+    currListItemIndex: 0,
+  };
